Move FakeMail test helper to tests.app.helpers
